Split up config module
